Bug fix & Add Synth.Dx7.Dx7Voice
Dx7Voice : MidiVoice is a special MidiVoice for Dx7 since the Dx7 voice is natively stereo.

Lfo for amplitude modulation in the Dx7Operator is still not working.
I tried.

Dx7Voice is slow.
It is impossible to use in real time.
Caching is also impressively slow.
The timber is pretty good.